+++ ytfetcher/cache/__init__.py
@@ -1,0 +1,3 @@
+from ytfetcher.cache.sqlite_cache import SQLiteCache
+
+__all__ = ["SQLiteCache"]