+++ CHANGELOG.md
@@ -6,10 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Added `convert_to_rows` utility method for converting `ChannelData` objects to Python dict for easily feed data to ML and RAG pipelines.
+- Added built-in cache strategy for fetching transcripts.
+- Added CLI argument for channel fetcher and playlist fetcher; `--all` argument now fetches ALL videos from a channel or playlist.
+- Added necessary tests for `PreviewRenderer` class.
 
 ### Changed
+- Changed `max_results` parameter to be optionally None which leads to fetch all videos from a channel if explicitly set.
+- Removed `timeout` parameter from `HTTPConfig` class.
+- Removed `httpx` library since it is unused.
 
 ### Fixed
+- Fixed `PreviewRenderer` fails if metadata values are None.
 
 ---